gateway: renamed/moved 'tcp::External' to 'connection::Handle'
Since we now have tcp-ipc-pair as a 'connection', the name _external_ is
misleading.

This patch moved the "external holder" to `group/connection.h´ with name
`group::connection::Holder` which is (arguable) a much better name.
